+++ CHANGELOG.md
@@ -8,6 +8,10 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Changed
+
+* Convert SSH urls into HTTPS urls in 'remote' command (#55)
+
 ## [0.8.0] - 2023-11-02
 
 ### Added