+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [0.9.0] - 2023-11-08
+
 ### Added
 
 * Display direct URL to zettels with `pyzet remote url <ID>` (#20).
@@ -120,7 +122,8 @@ Initial release.
 
 <!-- Links -->
 
-[Unreleased]: https://github.com/tpwo/pyzet/compare/v0.8.0...HEAD
+[Unreleased]: https://github.com/tpwo/pyzet/compare/v0.9.0...HEAD
+[0.9.0]: https://github.com/tpwo/pyzet/releases/tag/v0.9.0
 [0.8.0]: https://github.com/tpwo/pyzet/releases/tag/v0.8.0
 [0.7.0]: https://github.com/tpwo/pyzet/releases/tag/v0.7.0
 [0.6.0]: https://github.com/tpwo/pyzet/releases/tag/v0.6.0